Solve compile errors #15
