+++ boot/src/memory/segment/descriptor/table.rs
@@ -40,25 +40,6 @@ impl Table {
             .as_ptr() as u64
     }
 
-    pub fn continuous_free_descriptor_indices(&self, length: usize) -> Option<Range<usize>> {
-        let indices: Range<usize> = self.free_descriptor_indices()
-            .into_iter()
-            .fold(0..0, |indices, index| if indices.len() == length {
-                indices
-            } else if indices.end == index {
-                indices.start..index + 1
-            } else {
-                index..index + 1
-            });
-        (indices.len() == length).then_some(indices)
-    }
-
-    pub fn descriptor(&self, selector: &Selector) -> Option<Interface> {
-        self.index2descriptor()
-            .into_iter()
-            .find_map(|(index, interface)| (index == selector.get_index() as usize).then_some(interface))
-    }
-
     pub fn get() -> Self {
         Register::get().into()
     }
@@ -91,23 +72,6 @@ impl Table {
         let size: usize = length * mem::size_of::<short::Descriptor>();
         let limit: usize = size - 1;
         limit as u16
-    }
-
-    pub fn set_descriptor(&mut self, index: usize, descriptor: &Interface) {
-        if descriptor.is_long_descriptor() {
-            let descriptor: long::Descriptor = descriptor.into();
-            let descriptor: u128 = descriptor.into();
-            let lower_descriptor: u64 = (descriptor & ((1 << u64::BITS) - 1)) as u64;
-            let higher_descriptor: u64 = (descriptor >> u64::BITS) as u64;
-            self.descriptors[index] = lower_descriptor;
-            self.descriptors[index + 1] = higher_descriptor;
-        } else if descriptor.is_short_descriptor() {
-            let descriptor: short::Descriptor = descriptor.into();
-            let descriptor: u64 = descriptor.into();
-            self.descriptors[index] = descriptor;
-        } else {
-            panic!("Can't set a segment descriptor.");
-        }
     }
 
     fn free_descriptor_indices(&self) -> BTreeSet<usize> {
