+++ boot/src/memory/segment/selector.rs
@@ -17,13 +17,6 @@ pub struct Selector {
 }
 
 impl Selector {
-    pub fn create(index: u16, ti: bool, rpl: u8) -> Self {
-        Self::default()
-            .with_index(index)
-            .with_ti(ti)
-            .with_rpl(rpl)
-    }
-
     #[inline(never)]
     pub fn cs() -> Self {
         let cs: u16;
