+++ kernel/src/application.rs
@@ -1,0 +1,2 @@
+pub const PRIVILEGE_LEVEL: u8 = 3;
+
